feat: 스타일 수정, KeyType 파일 분리
