+++ assignment/2025-03-24/strong_box.dart
@@ -1,3 +1,5 @@
+import 'index.dart';
+
 class StrongBox<E> {
   E? _data;
   KeyType type;
@@ -17,26 +19,6 @@ class StrongBox<E> {
       return _data;
     } else {
       return null;
-    }
-  }
-}
-
-enum KeyType {
-  padlock,
-  button,
-  dial,
-  finger;
-
-  int get openCountLimit {
-    switch (this) {
-      case KeyType.padlock:
-        return 1_024;
-      case KeyType.button:
-        return 10_000;
-      case KeyType.dial:
-        return 30_000;
-      case KeyType.finger:
-        return 1_000_000;
     }
   }
 }
